api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
marshmallow==3.20.1
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
pytest==7.4.3
